[mypyc] Refactor: extract class-related genops to a new module
Also extract some utilities.
